specify tag/version

+++ update.sh
@@ -1,5 +1,13 @@
 #!/bin/sh
-git clone --depth=1 --branch=trunk git@github.com:apache/cassandra tmp
+
+TAG=$1
+
+if [[ "$TAG" == "" ]]; then
+  echo "no branch or tag specified"
+  exit 1
+fi
+
+git clone --depth=1 --branch=$TAG git@github.com:apache/cassandra tmp
 cp tmp/interface/cassandra.thrift cassandra-thrift/
 rm -Rf tmp
 thrift --gen php:oop,validate cassandra-thrift/cassandra.thrift
